+++ src/OpenGeoSysUncertaintyQuantification/convienence.jl
@@ -1,65 +1,138 @@
-AltInplaceOpsInterface.add!(a::Vector{Float64}, b::Vector{Float64}) = a .+= b
-AltInplaceOpsInterface.add!(a::Vector{Float64}, b::Float64) = a .+= b 
-mul!(a::Vector{Float64}, b::Float64) = a .*= b
-mul!(a::Vector{Float64}, b::Vector{Float64}) = a .*= b
-mul!(a::Vector{Float64}, b::Vector{Float64}, c::Float64) = begin
-	for i = 1:length(a)
+function AltInplaceOpsInterface.add!(a::Vector{Float64}, b::Vector{Float64})
+	@inbounds @simd for i in 1:length(a)
+		 a[i] += b[i]
+	end
+	return nothing
+end
+function AltInplaceOpsInterface.add!(a::Vector{Float64}, b::Float64)
+	@inbounds @simd for i in 1:length(a)
+		 a[i] += b
+	end
+	return nothing
+end
+function mul!(a::Vector{Float64}, b::Float64)
+	@inbounds @simd for i in 1:length(a)
+		 a[i] *= b
+	end
+	return nothing	
+end
+function mul!(a::Vector{Float64}, b::Vector{Float64})
+	@inbounds @simd for i in 1:length(a)
+		 a[i] *= b[i]
+	end
+	return nothing	
+end
+function mul!(a::Vector{Float64}, b::Vector{Float64}, c::Float64)
+	@inbounds @simd for i = 1:length(a)
 		a[i] = b[i]*c
 	end
 	return nothing
 end
-AltInplaceOpsInterface.minus!(a::Vector{Float64}, b::Vector{Float64}) = a .-= b
-AltInplaceOpsInterface.pow!(a::Vector{Float64}, b::Int64) = a .^= b
-AltInplaceOpsInterface.pow!(a::Vector{Float64}, b::Float64) = a .^= b
+function AltInplaceOpsInterface.minus!(a::Vector{Float64}, b::Vector{Float64})
+	@inbounds @simd for i in 1:length(a)
+		 a[i] -= b[i]
+	end
+	return nothing	
+end
+function AltInplaceOpsInterface.pow!(a::Vector{Float64}, b::Int64)
+	@inbounds @simd for i in 1:length(a)
+		 a[i] ^= b
+	end
+	return nothing		
+end
+function AltInplaceOpsInterface.pow!(a::Vector{Float64}, b::Float64)
+	@inbounds @simd for i in 1:length(a)
+		 a[i] ^= b
+	end
+	return nothing		
+end
 
 Base.fill!(a::Vector{Vector{Float64}}, b::Float64) = foreach(x->fill!(x,b), a)
 Base.similar(a::Vector{Vector{Float64}}) = map(x->similar(x),a)
 Base.zero(a::Vector{Vector{Float64}}) = map(zero,a)
 
 AltInplaceOpsInterface.add!(a::Vector{Vector{Float64}}, b::Vector{Vector{Float64}}) = begin 
-	for i = 1:length(a); add!(a[i],b[i]) end 
+	@inbounds for i = 1:length(a); add!(a[i],b[i]) end 
 	return nothing
 end
 AltInplaceOpsInterface.add!(a::Vector{Vector{Float64}}, b::Float64) = begin
-	for i = 1:length(a); add!(a[i],b) end
+	@inbounds for i = 1:length(a); add!(a[i],b) end
 	return nothing
 end
 mul!(a::Vector{Vector{Float64}}, b::Float64) = begin
-	for i = 1:length(a); mul!(a[i],b) end
+	@inbounds for i = 1:length(a); mul!(a[i],b) end
[…]
