### 2024-04-03 #6517 @baxmittens
@kthyng

I think I have fixed most of the typos, but I will have another good look at the paper before the review.

Here are five suggestions:

[rafaelorozco](https://github.com/rafaelorozco)
[boriskaus](https://github.com/boriskaus)
[ma-sadeghi](https://github.com/ma-sadeghi)
[ziyiyin97](https://github.com/ziyiyin97)
[jakelangham](https://github.com/jakelangham)
