### 2024-05-03 #6517 @kanishkan91
@peanutfun Are you interested in reviewing this manuscript for JOSS? Seems like a good fit given interests and previous reviews.
