### 2024-05-06 #6517 @kanishkan91
@dannys4 Would you be interested in reviewing this software package for the Journal of Open Source Software? You can find the pdf of the paper, GitHub repo in this thread above.
