### 2024-05-06 #6725 @editorialbot
[REVIEW]: OpenGeoSysUncertaintyQuantification.jl: a Julia library implementing an uncertainty quantification toolbox for OpenGeoSys
**Submitting author:** @baxmittens (<a href="http://orcid.org/0000-0001-9954-294X">maximilian bittens</a>)
**Repository:** https://github.com/baxmittens/OpenGeoSysUncertaintyQuantification.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.6
