### 2024-05-06 #6725 @kanishkan91
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1137/100786356 is OK
> - 10.1137/060663660 is OK
> - 10.1016/j.jcp.2009.01.006 is OK
> - 10.48550/arXiv.1509.01462 is OK
> - 10.21105/joss.04300 is OK
> - 10.1109/38.865875 is OK
> - 10.21105/joss.05003 is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: Quadrature and interpolation formulas for tensor p...
> - 10.1007/s12665-012-1546-x may be a valid DOI for title: OpenGeoSys: an open-source initiative for numerica...
>
> INVALID DOIs
>
> - None
> ```

@baxmittens, in the meantime see comment above. Looks like a DOI is missing for a reference. Could you get that corrected? Thanks!
