### 2024-05-06 #6725 @kanishkan91
@dannys4 - When you get a chance, could you fill out this form here- https://reviewers.joss.theoj.org/reviewers/new? This would convey JOSS editors your willingness to review for JOSS in the future based on interest. Let me know what you think. Thanks!
