### 2024-05-13 #6725 @dannys4
The software paper largely looks great and very clear, there's just a few (admittedly narrow) nitpicks I have:
- Distributions.jl should probably be properly cited in the referencessince it is referenced in the work (citation [here](https://github.com/JuliaStats/Distributions.jl/blob/master/CITATION.bib))
- Sobol' should have the apostrophe at the end for proper method attribution.

Less a comment and more of a question for @kanishkan91 : the paper states "special attention was paid to implement allocation-free in-place variants of all necessary math operators for all output datatypes such as `VTUFile` or `XDMF3File`". Is this a "performance claim" as suggested in the checklist and, if so, how rigorously should I try to check this?
