### 2024-05-13 #6725 @dannys4
On the examples side, I was having trouble getting the "plug and play" provided example to work.
- A minor detail is that the docs for `altered_StochasticOGSModelParams.xml` do not match the actual file in the test repo and I wasn't sure if I was supposed to change anything to make the mean and variance work properly.
- I also wasn't sure if I was supposed to be able to generate the plots provided in the documentation (and if so, how).
- It only started working for me once I used the version of `DistributedSparseGrids` on the `main` branch. It seems like it was known that `PlotlyJS` has some issues, and this was preventing my functionality (Julia 1.10.3)
- Once I got that figured out, it looks like there's some pathing issue going on. I get to `start!(ogsuqasg)` and then I get a bunch of output like
```
julia> start!(ogsuqasg)
[ Info: Starting 65 simulatoin calls
      From worker 2:    1_1_1_1Progress: 100%|███████████████████████████████████████████████████████████████████████████████████████████████| Time: 0:00:00

      From worker 3:	1_2_1_1
      From worker 2:	ogs call @[0.0, 0.0]
      From worker 3:	ogs call @[0.0, -1.0]
...
ERROR: TaskFailedException

    nested task error: On worker 2:
    IOError: could not spawn `/path/to/ogs/bin/ogs -o ./Res/1_1_1_1 ./Res/1_1_1_1/point_heat_source_2D.prj`: no such file or directory (ENOENT)
    Stacktrace:
      [1] _spawn_primitive
        @ ./process.jl:128
      [2] #784
        @ ./process.jl:139 [inlined]
      [3] setup_stdios
        @ ./process.jl:223
      [4] _spawn
        @ ./process.jl:138 [inlined]
      [5] _spawn
        @ ./process.jl:166
      [6] #run#798
        @ ./process.jl:479
      [7] run
        @ ./process.jl:477 [inlined]
      [8] fun
        @ ~/.julia/dev/OpenGeoSysUncertaintyQuantification/test/ex1/user_functions.jl:46
      [9] fun
        @ ~/.julia/dev/OpenGeoSysUncertaintyQuantification/test/ex1/user_functions.jl:41
     [10] #invokelatest#2
        @ ./essentials.jl:892
     [11] invokelatest
        @ ./essentials.jl:889
     [12] #110
        @ ~/.julia/juliaup/julia-1.10.3+0.aarch64.apple.darwin14/share/julia/stdlib/v1.10/Distributed/src/process_messages.jl:287
     [13] run_work_thunk
        @ ~/.julia/juliaup/julia-1.10.3+0.aarch64.apple.darwin14/share/julia/stdlib/v1.10/Distributed/src/process_messages.jl:70
     [14] #109
        @ ~/.julia/juliaup/julia-1.10.3+0.aarch64.apple.darwin14/share/julia/stdlib/v1.10/Distributed/src/process_messages.jl:287
    Stacktrace:
     [1] remotecall_fetch(::Function, ::Distributed.Worker, ::StaticArraysCore.SVector{…}, ::Vararg{…}; kwargs::@Kwargs{})
       @ Distributed ~/.julia/juliaup/julia-1.10.3+0.aarch64.apple.darwin14/share/julia/stdlib/v1.10/Distributed/src/remotecall.jl:465
     [2] remotecall_fetch(::Function, ::Distributed.Worker, ::StaticArraysCore.SVector{2, Float64}, ::Vararg{Any})
       @ Distributed ~/.julia/juliaup/julia-1.10.3+0.aarch64.apple.darwin14/share/julia/stdlib/v1.10/Distributed/src/remotecall.jl:454
     [3] remotecall_fetch(::Function, ::Int64, ::StaticArraysCore.SVector{2, Float64}, ::Vararg{Any}; kwargs::@Kwargs{})
       @ Distributed ~/.julia/juliaup/julia-1.10.3+0.aarch64.apple.darwin14/share/julia/stdlib/v1.10/Distributed/src/remotecall.jl:492
     [4] remotecall_pool(::Function, ::Function, ::Distributed.WorkerPool, ::StaticArraysCore.SVector{…}, ::Vararg{…}; kwargs::@Kwargs{})
       @ Distributed ~/.julia/juliaup/julia-1.10.3+0.aarch64.apple.darwin14/share/julia/stdlib/v1.10/Distributed/src/workerpool.jl:126
     [5] remotecall_pool
       @ ~/.julia/juliaup/julia-1.10.3+0.aarch64.apple.darwin14/share/julia/stdlib/v1.10/Distributed/src/workerpool.jl:123 [inlined]
     [6] remotecall_fetch
       @ ~/.julia/juliaup/julia-1.10.3+0.aarch64.apple.darwin14/share/julia/stdlib/v1.10/Distributed/src/workerpool.jl:232 [inlined]
     [7] (::DistributedSparseGrids.var"#41#42"{…})()
       @ DistributedSparseGrids ~/.julia/dev/DistributedSparseGrids/src/DistributedSparseGrids.jl:353

...and 64 more exceptions.

Stacktrace:
 [1] sync_end(c::Channel{Any})
   @ Base ./task.jl:448
 [2] macro expansion
   @ ./task.jl:480 [inlined]
 [3] distributed_fvals!(asg::AdaptiveHierarchicalSparseGrid{…}, cpts::Vector{…}, fun::typeof(fun), worker_ids::Vector{…})
   @ DistributedSparseGrids ~/.julia/dev/DistributedSparseGrids/src/DistributedSparseGrids.jl:344
 [4] distributed_init_weights_inplace_ops!(asg::AdaptiveHierarchicalSparseGrid{…}, cpts::Vector{…}, fun::typeof(fun), worker_ids::Vector{…})
   @ DistributedSparseGrids ~/.julia/dev/DistributedSparseGrids/src/DistributedSparseGrids.jl:423
 [5] macro expansion
   @ ./timing.jl:279 [inlined]
 [6] start!(ogsuqasg::OGSUQASG, refinetohypercube::Bool)
   @ OpenGeoSysUncertaintyQuantification ~/.julia/dev/OpenGeoSysUncertaintyQuantification/src/OpenGeoSysUncertaintyQuantification.jl:487
 [7] start!(ogsuqasg::OGSUQASG)
   @ OpenGeoSysUncertaintyQuantification ~/.julia/dev/OpenGeoSysUncertaintyQuantification/src/OpenGeoSysUncertaintyQuantification.jl:473
 [8] top-level scope
   @ REPL[18]:1
Some type information was truncated. Use `show(err)` to see complete types.
```
where I get
```
shell> ls Res/
1_1_1_1		1_5_1_10	2_1_3_1		2_4_1_2		3_1_4_1		4_1_2_1		4_2_6_3		5_1_6_1
1_2_1_1		1_5_1_12	2_2_1_1		2_4_1_4		3_2_2_1		4_1_4_1		4_2_8_1		5_1_8_1
1_2_1_3		1_5_1_14	2_2_1_3		2_4_1_6		3_2_2_3		4_1_6_1		4_2_8_3
1_3_1_2		1_5_1_16	2_2_3_1		2_4_1_8		3_2_4_1		4_1_8_1		5_1_10_1
1_3_1_4		1_5_1_2		2_2_3_3		2_4_3_2		3_2_4_3		4_2_2_1		5_1_12_1
1_4_1_2		1_5_1_4		2_3_1_2		2_4_3_4		3_3_2_2		4_2_2_3		5_1_14_1
1_4_1_4		1_5_1_6		2_3_1_4		2_4_3_6		3_3_2_4		4_2_4_1		5_1_16_1
1_4_1_6		1_5_1_8		2_3_3_2		2_4_3_8		3_3_4_2		4_2_4_3		5_1_2_1
1_4_1_8		2_1_1_1		2_3_3_4		3_1_2_1		3_3_4_4		4_2_6_1		5_1_4_1
```
