### 2024-05-14 #6725 @baxmittens
> The software paper largely looks great and very clear, there's just a few (admittedly narrow) nitpicks I have:
>
> * Distributions.jl should probably be properly cited in the referencessince it is referenced in the work (citation [here](https://github.com/JuliaStats/Distributions.jl/blob/master/CITATION.bib))
> * Sobol' should have the apostrophe at the end for proper method attribution.
>
> Less a comment and more of a question for @kanishkan91 : the paper states "special attention was paid to implement allocation-free in-place variants of all necessary math operators for all output datatypes such as `VTUFile` or `XDMF3File`". Is this a "performance claim" as suggested in the checklist and, if so, how rigorously should I try to check this?

Hello @dannys4

thank you for reviewing this software. I appreciate it.
I will go to your comments and alter my code appropriately. If I fix the things accordingly, I will write here again.

I added the citation and apostrophe you've suggested.

The `performance claim` can be tested likewise:

```julia
import AltInplaceOpsInterface: add!
import LinearAlgebra: mul!
using BenchmarkTools
xdmf = XDMF3File("./Res/1_1_1_1/PointHeatSource_quarter_002_2nd.xdmf");
tmp = deepcopy(xdmf);
@btime $xdmf + $xdmf;
# 878.125 μs (1472 allocations: 3.13 MiB)
@btime add!($xdmf, $tmp)
# 164.500 μs (0 allocations: 0 bytes)
@btime $xdmf * $xdmf;
# 599.958 μs (1472 allocations: 3.13 MiB)
@btime mul!($xdmf, $tmp, $tmp)
# 60.792 μs (0 allocations: 0 bytes)
```

At the moment, I use a nasty thing called [AltInplaceOpsInterface.jl](https://github.com/baxmittens/AltInplaceOpsInterface.jl) for in-place operation since Julia does not seem to support broadcasting for nested data types and may never will, see e.g. https://discourse.julialang.org/t/efficient-weighted-sum-for-arbitrary-data-types/113992
