### 2024-05-14 f17dbb3
update paper and examples
