+++ paper.md
@@ -67,13 +67,13 @@ Features of OpenGeoSysUncertaintyQuantification.jl:
 
 - Setup of stochastic OGS6 projects (see [docs](https://baxmittens.github.io/OpenGeoSysUncertaintyQuantification.jl/dev/)).
 
-- Definition of input parameters and (truncated) input distributions via [Distributions.jl](https://github.com/JuliaStats/Distributions.jl).
+- Definition of input parameters and (truncated) input distributions via  [Distributions.jl](https://github.com/JuliaStats/Distributions.jl) [Distributions.jl-2019].
 
 - Adaptive sparse grid surrogate modeling of the physical state space. All snapshots have to fit in the system memory.
 
 - Distributed Monte Carlo integration. Snapshots do not have to fit in the system memory.
 
-- Monte Carlo integrated Sobol indices.
+- Monte Carlo integrated Sobol' indices.
 
 - Monte Carlo or Latin Hypercube integrated Morris means.
 
