### 2024-05-14 #6725 @kanishkan91
@dannys4 , @baxmittens looks like you guys are communicating through the issues, which is great!

@dannys4 On your question above, if a specific claim is made in the paper regarding processing for a specific data type, there may be two things to check-

1. If the test cases cover those special cases
2. If there are specific examples provided in the documentation and vignettes/tutorials.

But again seems like @baxmittens responded to that already!
