### 2024-05-15 #6725 @baxmittens
> * A minor detail is that the docs for `altered_StochasticOGSModelParams.xml` do not match the actual file in the test repo and I wasn't sure if I was supposed to change anything to make the mean and variance to match the documented example.

I fixed that

> * I also wasn't sure if I was supposed to be able to generate the plots provided in the documentation (and if so, how).

I updated the `./test/ex1/start.jl`

in line 22-30, you can find

```julia
# import Pkg
# Pkg.add("PlotlyJS")
# Pkg.add("DistributedSparseGridsPlotting")

using PlotlyJS
using DistributedSparseGridsPlotting
display(PlotlyJS.plot([PlotlyJS.scatter3d(ogsuqasg.asg), surface_inplace_ops(ogsuqasg.asg, 20, x->maximum(x["temperature_interpolated"]))], PlotlyJS.Layout(title="ASG-Surrogate response function ASG(x) - max. temp")))
display(PlotlyJS.plot([PlotlyJS.scatter3d(asg_expval), surface_inplace_ops(asg_expval, 50, x->maximum(x["temperature_interpolated"]))], PlotlyJS.Layout(title="ASG(x)*pdf(x) - max. temp")))
display(PlotlyJS.plot([PlotlyJS.scatter3d(asg_varval), surface_inplace_ops(asg_varval, 50, x->maximum(x["temperature_interpolated"]))], PlotlyJS.Layout(title="(ASG(x)-𝔼(x))^2*pdf(x) - max. temp")))
```

> * It only started working for me once I used the version of `DistributedSparseGrids` on the `main` branch. It seems like it was known that `PlotlyJS` has some issues, and this was preventing my functionality (Julia 1.10.3). This may be above my paygrade as a reviewer, but there's definitely `compat`s missing for the dependencies that would prevent some of these issues.

I agree. Recently, PlotlyJS stopped working (at least for me) if it is included in remote workers.
I did throw PlotlyJS out since I plan to switch to Makie, and, in addition, I do not like to have these heavy graphic libraries as a dependency, anyway. I issued a new version. this should be fixed, I hope.

> * Once I got that figured out, it looks like there's some pathing issue going on. I get to `start!(ogsuqasg)` and then I get a bunch of output like

You have to change `simcall` from `/path/to/ogs/bin/ogs` to the `path` of you ogs installation in the `altered_StochasticOGSModelParams.xml`
Installing OGS can be a bit cumbersome. I can provide some help if you need it.
