+++ test/ex1/generate_stoch_model.jl
@@ -1,14 +1,15 @@
 using OpenGeoSysUncertaintyQuantification
 
 projectfile="./project/point_heat_source_2D.prj"
-simcall="/home/ogs_auto_jenkins/temporary_versions/native/master/ogs6_2023-02-23/bin/ogs"
+simcall="ogs" # ogs binary has to be in path. otherwise insert your "path/to/ogs"
 additionalprojecfilespath="./mesh"
 outputpath="./Res"
-postprocfiles=["PointHeatSource_ts_10_t_50000.000000.vtu"]
+postprocfiles=["PointHeatSource_quarter_002_2nd.xdmf"]
 outputpath="./Res"
 stochmethod=AdaptiveHierarchicalSparseGrid
+n_workers = 10
 
-stochparampathes = loadStochasticParameters("altered_PossibleStochasticParameters.xml")
+stochparampathes = loadStochasticParameters("StochasticParameters.xml")
 	
 stochasticmodelparams = generateStochasticOGSModell(
 	projectfile,
@@ -17,6 +18,26 @@ stochasticmodelparams = generateStochasticOGSModell(
 	postprocfiles,
 	stochparampathes,
 	outputpath,
-	stochmethod)
+	stochmethod,
+	n_workers)
+
+# alter the stochastic parameters
+stoch_params = stoch_parameters(stochasticmodelparams)
+@assert contains(stoch_params[1].path, "AqueousLiquid")
+stoch_params[1].dist = Normal(0.6,0.175)
+stoch_params[1].lower_bound = 0.3
+stoch_params[1].upper_bound = 0.9
+stoch_params[2].dist = Normal(0.45,0.15)
+stoch_params[2].lower_bound = 0.1
+stoch_params[2].upper_bound = 0.8
+
+write(stochasticmodelparams)
 
 samplemethodparams = generateSampleMethodModel(stochasticmodelparams)
+
+#alter sample method params
+samplemethodparams.init_lvl = 4
+samplemethodparams.maxlvl = 12
+samplemethodparams.tol = 0.025
+
+write(samplemethodparams)
