+++ src/OpenGeoSysUncertaintyQuantification.jl
@@ -622,6 +622,6 @@ export OGS6ProjectParams, StochasticOGS6Parameter, StochasticOGSModelParams, Sam
 	setStochasticParameters!, setStochasticParameter!, lin_func, CPtoStoch, StochtoCP, pdf, getElementbyPath, XDMF3File, XDMFData, MonteCarlo, MonteCarloSobol, MonteCarloMorris, MonteCarloMorrisParams,
 	variance, 𝔼, XMLFile, XML2Julia, init, start!, integrate_nodal_result, integrate_cell_result, integrate_area, integrate_result, ogs6_modeldef, stoch_parameters, scalar_sobolindex_from_multifield_result,
 	scalar_sobolindex_from_field_result, scalar_sobolindex_from_multifield_result, write_sobol_multifield_result_to_XDMF, write_sobol_field_result_to_XDMF, dependend_tensor_parameter!, dependend_parameter!,
-	empirical_cdf
+	empirical_cdf, writeStochasticParameters, writeStochasticOGSModelParams
 
 end # module
