+++ docs/src/index.md
@@ -42,9 +42,9 @@ creates an xml-file which defines the so-called `StochasticOGSModelParams`. It i
 - the `simcall` (e.g. `"path/to/ogs/bin/ogs"`), 
 - a `additionalprojecfilespath` where meshes and other files can be located which are copied in each individual folder for a OGS6-snapshot, 
 - the path to one or more `postprocfile`s, 
-- the stochpathes, generated with `generatePossibleStochasticParameters`, manipulated by the user, and loaded by the `loadStochasticParameters`-function,
+- the stochpathes, generated with `generatePossibleStochasticParameters`. Can be manipulated by the user and loaded by the `loadStochasticParameters`-function,
 - an `outputpath`, where all snapshots will be stored,
-- a `stochmethod` (Sparse grid or Monte-Carlo, where Monte-Carlo is not yet implemented),
+- a `stochmethod` (Sparse grid or Monte-Carlo, see [Library](# Library)),
 - the number of local workers `n_local_workers`, and, 
 - the filename `sogsfile` under which the model is stored as an xml-file. 
 
@@ -138,7 +138,7 @@ write(stochasticmodelparams)
 samplemethodparams = generateSampleMethodModel(stochasticmodelparams)
 
 #alter sample method params
-samplemethodparams.RT = XDMFData #according to `fun` user_functions.jl
+samplemethodparams.RT = XDMF3File #according to `fun` user_functions.jl
 samplemethodparams.init_lvl = 4
 samplemethodparams.maxlvl = 12
 samplemethodparams.tol = 0.025
@@ -184,7 +184,7 @@ ogsuqasg = init(ogsuqparams)
 start!(ogsuqasg)
 ```
 
-load the parameters `ogsuqparams`, initializes the model `ogsuqasg`, and, starts the sampling procedure. Finally the expected value is integrated.
+load the parameters `ogsuqparams`, initializes the model `ogsuqasg`, and, starts the sampling procedure. Finally the expected value is integrated. Note that since the pathes are relative, the `start!` function has to be called from the working directory.
 
 * Initializing the model `OpenGeoSysUncertaintyQuantification.init(ogsuqparams)` consists of two steps
 	
@@ -197,7 +197,7 @@ load the parameters `ogsuqparams`, initializes the model `ogsuqasg`, and, starts
 <table border="0"><tr>
 <td> 
 	<figure>
-		<img src="https://user-images.githubusercontent.com/100423479/223154558-4b94d7a2-e93b-45ef-9783-11437ae23b35.png" width="350" height="300" /><br>
+		<img src="./assets/asg.png" width="350" height="300" /><br>
 		<figcaption><em>resulting sparse grid</em></figcaption>
 	</figure>
 </td>
@@ -236,7 +236,7 @@ Hereby, the physical surrogate model, generated by the sampling of the model, is
 </tr></table>
 ```
 
-By integrating over the domain the expected value is computed. Below the pressure field and the temperature field are shown.
+By integrating over the stochstic domain, the expected value is computed. Below the pressure field and the temperature field are shown.
 ```@raw html
 <table border="0"><tr>
 <td> 
@@ -259,12 +259,11 @@ By integrating over the domain the expected value is computed. Below the pressur
 The variance can be computed by:
 
 ```julia
-varval,asg_varval = variance(ogsuqasg,expval);
-VTUFileHandler.rename!(varval,"varval_heatpointsource.vtu")
-write(varval)
+varval,asg_varval = variance(ogsuqasg, expval)
+write(varval, "varval.xdmf", "varval.h5")
 ```
 
[…]
