+++ docs/src/index.md
@@ -69,6 +69,10 @@ creates an xml-file `anafile` with all necessary parameters for the chosen sampl
 
 In this chapter, [Example 1](https://github.com/baxmittens/OpenGeoSysUncertaintyQuantification.jl/tree/main/test/ex1) is used to illustrate the workflow. The underlying deterministic OGS6 project is the [point heat source example](https://www.opengeosys.org/docs/benchmarks/th2m/saturatedpointheatsource/) ([Thermo-Richards-Mechanics project files](https://gitlab.opengeosys.org/ogs/ogs/-/tree/master/Tests/Data/ThermoRichardsMechanics/PointHeatSource)).
 
+The complete example can be run by directing to the path `path/to/OpenGeoSysUncertaintyQuantification/test/ex1`, running Julia (if Julia is started by, e.g., `julia -t 10`, the adaptive sparse grid can take advantage of the extra threads) and starting the example by
+```julia
+include("run_ex1.jl")
+``` 
 
 ### Defining the stochastic dimensions
 
