### 2024-05-23 #6725 @dannys4
I'm trying to finish up the rest of this review as I'm able, but I feel like I'm pretty close. A few points of note:
> @baxmittens : I updated the documentation, too.

The new usage example looks and feels much smoother overall, and I think will work well. The only point of criticism I have is that at one point you require `DistributedSparseGridsPlotting` in `start.jl` instead of `DistributedSparseGrids` (or you `use` the right one, but don't import the right keywords for `surface_inplace_ops`, which causes an error. Otherwise, once I fix this, it seems to at least work on `main`.

The last two points on the review are on documentation/testing:
- There's plenty of documentation of how to write these `xdmf` files to memory, but I couldn't figure out from these docs how to actually load things back in. If I want to serialize the mean (or the sparse grid!) I get from this and use it in future computations, I feel like this would be not just helpful, but vital.
- I'm a little bit queasy on the testing side of things; there are a lot of keywords exported and very few called in the tests. Additionally, since one of them depends on the now-destroyed `ex1/altered...` files, the tests don't seem to run as written on `main`, and it doesn't look like the github workflow automatically runs these tests. I'm not a stickler on the last point (since running the tests are very easy in Julia), but they should probably be adjusted and, if possible, expanded on to include the functions that wouldn't be tested in the direct dependencies (e.g., ignoring the integration and interpolations presumably tested in `DistributedSparseGrids`).

Once these get addressed and the paper is rerendered, I'll test on the general registry version, reread and consider the review done.

As a remark that in no way influences my reviewing, I might suggest that you look into working with extensions in Julia as a way to work around the problem with plotting dependencies; it's not a perfect solution, but it's the best solution I know of for handling these issues.
