### 2024-05-24 #6725 @baxmittens
Hey @dannys4

thank you again for your thorough review. I will work on fixing things as soon as possible, which will be mid of next week I think as i am traveling for a long weekend.

> * There's plenty of documentation of how to write these `xdmf` files to memory, but I couldn't figure out from these docs how to actually load things back in. If I want to serialize the mean (or the sparse grid!) I get from this and use it in future computations, I feel like this would be not just helpful, but vital.

I can read xdmf-files by

```julia
xdmf = XDMFFileHandler.XDMF3File("path/to/xdmf.xdmf")
```

I'll will add a description. The use of xdmf instead of vtu is fairly new in OpenGeoSys.
I am afraid, I don't really get the rest of your point. I don't get why I should want to serialize the mean or the sparse grid. Let me try to explain, how it is meant to be.

The first time `start.jl` is run, the `Res` folder gets populated, according to the experimental design of a sparse grid or some other method, with individual realizations of OGS projects at their specific location in the state space for which ogs is called in a further step and the postprocessing results are generated and also saved in the individual folders with the project data. Stochastic postprocessing can be conducted if all samples are generated and loaded into memory
The next time `start.jl` is run, OGS will not be called once again for each realization (if it is not parameterized differently in the `user_functions.jl` file), and only the postprocessing results will be loaded into memory, which is done fairly fast. So, if I need a result, I can load it from an xdmf file; if I need the sparse grid in memory for generating more post-processing results, I will just run the `start.jl` once again.
Does that make sense? Or did I get you wrong?

> * I'm a little bit queasy on the testing side of things; there are a lot of keywords exported and very few called in the tests. Additionally, since one of them depends on the now-destroyed `ex1/altered...` files, the tests don't seem to run as written on `main`, and it doesn't look like the github workflow automatically runs these tests. I'm not a stickler on the last point (since running the tests are very easy in Julia), but they should probably be adjusted and, if possible, expanded on to include the functions that wouldn't be tested in the direct dependencies (e.g., ignoring the integration and interpolations presumably tested in `DistributedSparseGrids`).

I did not plan to run the examples in the tests since this would require a OGS binary. I will think about how to resolve this issue.

> As a remark that in no way influences my reviewing, I might suggest that you look into working with extensions in Julia as a way to work around the problem with plotting dependencies; it's not a perfect solution, but it's the best solution I know of for handling these issues.

Thank you for your suggestion. I will have a look into it. Maybe this fixes things for me.
