### 2024-05-24 #6725 @dannys4
> I'll will add a description. The use of xdmf instead of vtu is fairly new in OpenGeoSys.

Sounds great!

> Let me try to explain, how it is meant to be.

I see. I think I was a little confused about the end-use and this explains it very well.

> I did not plan to run the examples in the tests since this would require a OGS binary. I will think about how to resolve this issue.

I don't really see a good solution here; I don't want to get too involved in the nitty-gritty software design, but is there any way you can extract the binary interface such that you can test that it has the correct behavior with output from a dummy OGS process (i.e., record OGS output from a "reference" session that you know works and somehow use that to test the behavior of how OGSUQ.jl responds)? I understand if this ends up not being viable, it's just the main point of contention when finishing the checklist.
