### 2024-05-28 baxmittens/OpenGeoSysUncertaintyQuantification.jl#3 @ziyiyin97
JOSS review
https://github.com/openjournals/joss-reviews/issues/6725

Apart from comments from the other reviewer, I have the following comments about the manuscript itself:

1. The first paragraph in the statement of need can benefit from a reference.
2. Line 26-29 is a bit confusing. Probably can rephrase to "This package was designed with the aim to be explicitly related to the OGS6 project. However, care was taken to ensure that the underlying packages could potentially be still applicable to other projects."
3. Line 38: it would be good if authors can briefly indicate their contribution on top of other relevant Julia packages.

I will post more comments for the software in a bit ...
