+++ .github/workflows/CI.yml
@@ -10,8 +10,7 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.7'
-          - '1.8.5'
+          - '1.10.2'
         os:
           - ubuntu-latest
         arch:
