+++ test/install_ogs.sh
@@ -1,3 +1,4 @@
+# install ogs via pip
 # only works with python versions < 3.12 (e.g. 3.11.9)
 mkdir ./ogspyvenv
 python3 -m venv ./ogspyvenv
