+++ test/test_project_setup.jl
@@ -1,5 +1,5 @@
-file_stoch_model = "./ex1/altered_StochasticOGSModelParams.xml"
-file_sample_params = "./ex1/altered_SampleMethodParams.xml"
+file_stoch_model = "./ex1/StochasticOGSModelParams.xml"
+file_sample_params = "./ex1/SampleMethodParams.xml"
 
 ogsuqparams = OGSUQParams(file_stoch_model, file_sample_params)
 
