+++ test/runtests.jl
@@ -1,5 +1,7 @@
 using OpenGeoSysUncertaintyQuantification
 using Test
+import Pkg
+Pkg.develop("OpenGeoSysUncertaintyQuantification")
 
 @testset "Project setup" begin
     include("test_project_setup.jl")
@@ -11,8 +13,7 @@ end
 
 @testset "OGS run" begin
     println(readdir())
-    prj_dir = joinpath(@__DIR__,"ex1")
-    cd(prj_dir) #has to be called from project directory
+    cd("ex1") #has to be called from project directory
     println(readdir())
     include("run_ex1.jl")
 end
