+++ test/ex1/generate_stoch_model.jl
@@ -1,20 +1,25 @@
 using OpenGeoSysUncertaintyQuantification
 
-projectfile="./project/point_heat_source_2D.prj"
+__relpath__ = relpath(@__DIR__, "./")
+projectfile= joinpath(__relpath__,"project","point_heat_source_2D.prj")
+user_functions_file = joinpath(__relpath__, "user_functions.jl")
+output_xml = joinpath(__relpath__, "StochasticOGSModelParams.xml")
+stoch_params_xml = joinpath(__relpath__, "StochasticParameters.xml")
+samplemethod_output_xml = joinpath(__relpath__, "SampleMethodParams.xml")
+
 simcall="ogs" # ogs binary has to be in path. otherwise insert your "path/to/ogs"
 if haskey(ENV, "OGS_BINARY")
 	@info "using $(ENV["OGS_BINARY"]) as binary"
 	simcall = ENV["OGS_BINARY"]
 end
-additionalprojecfilespath="./mesh"
-outputpath="./Res"
+additionalprojecfilespath=joinpath(__relpath__,"mesh")
+outputpath=joinpath(__relpath__,"Res")
 postprocfiles=["PointHeatSource_quarter_002_2nd.xdmf"]
-outputpath="./Res"
 stochmethod=AdaptiveHierarchicalSparseGrid
 n_workers = 4
 
-stochparampathes = loadStochasticParameters("StochasticParameters.xml")
-	
+stochparampathes = loadStochasticParameters(stoch_params_xml)
+
 stochasticmodelparams = generateStochasticOGSModell(
 	projectfile,
 	simcall,
@@ -23,7 +28,9 @@ stochasticmodelparams = generateStochasticOGSModell(
 	stochparampathes,
 	outputpath,
 	stochmethod,
-	n_workers)
+	n_workers,
+	user_functions_file,
+	output_xml)
 
 # alter the stochastic parameters
 stoch_params = stoch_parameters(stochasticmodelparams)
@@ -36,8 +43,7 @@ stoch_params[2].lower_bound = 0.1
 stoch_params[2].upper_bound = 0.8
 
 write(stochasticmodelparams)
-
-samplemethodparams = generateSampleMethodModel(stochasticmodelparams)
+samplemethodparams = generateSampleMethodModel(stochasticmodelparams, samplemethod_output_xml)
 
 #alter sample method params
 samplemethodparams.init_lvl = 4
