+++ test/ex1/generate_stoch_params_file.jl
@@ -1,9 +1,12 @@
 using OpenGeoSysUncertaintyQuantification
 
-projectfile="./project/point_heat_source_2D.prj"
+__relpath__ = relpath(@__DIR__, "./")
+projectfile = joinpath(__relpath__, "project", "point_heat_source_2D.prj")
+output_xml = joinpath(__relpath__, "StochasticParameters.xml")
+
 pathes = generatePossibleStochasticParameters(projectfile)
 
 porosity_ind = findfirst(x->contains(x,"@id/0") && contains(x,"porosity"), pathes)
 liquid_th_cond_ind = findfirst(x->contains(x,"@id/0") && contains(x,"AqueousLiquid") && contains(x,"thermal_conductivity"), pathes)
 
-writeStochasticParameters(pathes[[porosity_ind, liquid_th_cond_ind]], "./StochasticParameters.xml")
+writeStochasticParameters(pathes[[porosity_ind, liquid_th_cond_ind]], output_xml)
