+++ Project.toml
@@ -19,6 +19,7 @@ DistributedMonteCarlo = "1911db96-88a4-4c72-b932-f455b4e9470b"
 Format = "1fa38f19-a742-5d3f-a2b9-30dd87b9d5f8"
 PGFPlotsX = "8314cec4-20b6-5062-9cdb-752b83310925"
 PrettyTables = "08abe8d2-0d0c-5749-adfa-8a2ac140af0d"
+DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 
 [compat]
 julia = "1"
@@ -37,3 +38,4 @@ LinearAlgebra = "1"
 Format = "1"
 PGFPlotsX = "1"
 PrettyTables = "2"
+DelimitedFiles = "1"
