### 2024-05-29 6988aed
Update generate_stoch_model.jl
