### 2024-05-29 eacc72f
Update generate_stoch_model.jl
