### 2024-05-29 8dc0c79
Update generate_stoch_model.jl
