### 2024-05-29 c2ea70d
Update generate_stoch_model.jl
