+++ test/ex1/generate_stoch_model.jl
@@ -17,6 +17,7 @@ elseif isfile(joinpath(PATH, "../ogspyvenv/bin/ogs")) # for GitHub Action / test
 else
 	@info "using ogs as binary"
 end
+@info readdir(joinpath(PATH,".."))
 additionalprojecfilespath=joinpath(PATH,"mesh")
 outputpath=joinpath(PATH,"Res")
 postprocfiles=["PointHeatSource_quarter_002_2nd.xdmf"]
