### 2024-05-29 af24200
Update generate_stoch_model.jl
