### 2024-05-29 23eb6c4
Update generate_stoch_model.jl
