+++ test/ex1/generate_stoch_model.jl
@@ -11,15 +11,15 @@ simcall = "ogs" # ogs binary is in path, otherwise put your path/to/bin/ogs here
 if haskey(ENV, "OGS_BINARY") # installed with install_ogs.sh
 	@info "using $(ENV["OGS_BINARY"]) as binary"
 	simcall = ENV["OGS_BINARY"]
-elseif isfile(joinpath(PATH, "../ogspyvenv/bin/ogs")) # for GitHub Action / testing
-	@info "using ./ogspyvenv/bin/ogs as binary"
-	simcall = "./ogspyvenv/bin/ogs"
+elseif isfile(joinpath(PATH, "../../ogspyvenv/bin/ogs")) # for GitHub Action / testing
+	@info "using ../../ogspyvenv/bin/ogs as binary"
+	simcall = joinpath(PATH, "../../ogspyvenv/bin/ogs")
 else
 	@info "using ogs as binary"
 end
-@info readdir(joinpath(PATH,".."))
-@info readdir(joinpath(PATH,"..",".."))
-@info readdir(joinpath(PATH,"..","..",".."))
+#@info readdir(joinpath(PATH,".."))
+#@info readdir(joinpath(PATH,"..",".."))
+#@info readdir(joinpath(PATH,"..","..",".."))
 additionalprojecfilespath=joinpath(PATH,"mesh")
 outputpath=joinpath(PATH,"Res")
 postprocfiles=["PointHeatSource_quarter_002_2nd.xdmf"]
