+++ .github/workflows/CI.yml
@@ -31,7 +31,7 @@ jobs:
             ${{ runner.os }}-test-${{ env.cache-name }}-
             ${{ runner.os }}-test-
             ${{ runner.os }}-
-      - run: bash ./test/install_ogs.sh
+      - run: bash ./test/install_ogs.sh ./
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
       - uses: julia-actions/julia-processcoverage@v1
