+++ src/OpenGeoSysUncertaintyQuantification/utils.jl
@@ -304,8 +304,8 @@ function pdf(stoparams::Vector{StochasticOGS6Parameter}, x)
 end
 
 function install_ogs()
-	warning("python version < 3.12 has to be installed for this to work")
-	PATH = joinpath(splitpath(@__FILE__)[1:end-2]..., "test")
+	@warn "python version < 3.12 has to be installed for this to work"
+	PATH = joinpath(splitpath(@__FILE__)[1:end-3]..., "test")
 	installscript = joinpath(PATH, "install_ogs.sh")
 	run(`bash $installscript $PATH`)
 	return joinpath(PATH,"/ogspyvenv/bin/ogs")
