+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"lib/lib/#Library","page":"Library","title":"Library","text":"","category":"section"},{"location":"lib/lib/#Contents","page":"Library","title":"Contents","text":"","category":"section"},{"location":"lib/lib/","page":"Library","title":"Library","text":"Pages = [\"lib.md\"]\nDepth = 4","category":"page"},{"location":"lib/lib/#Project-setup","page":"Library","title":"Project setup","text":"","category":"section"},{"location":"lib/lib/","page":"Library","title":"Library","text":"OpenGeoSysUncertaintyQuantification.jl provides methods for the generation of a stochastic OGS6 model from a validly configured OGS6 project. generatePossibleStochasticParameters scans existing OGS6 projects for possible stochastic parameters. generateStochasticOGSModell is a helper function which creates an data structure containing all parameters describing the stochastic OGS model, defined by StochasticOGSModelParams. generateSampleMethodModel generates the data structure for the sample method, defined by the abstract type SampleMethodParams. Currently implented are: SparseGridParams, MonteCarloParams, MonteCarloSobolParams, or MonteCarloMorrisParams.","category":"page"},{"location":"lib/lib/","page":"Library","title":"Library","text":"generatePossibleStochasticParameters\nloadStochasticParameters\ngenerateStochasticOGSModell\ngenerateSampleMethodModel\nwriteStochasticParameters\nBase.write(::StochasticOGSModelParams)\nBase.write(::SampleMethodParams)","category":"page"},{"location":"lib/lib/#OpenGeoSysUncertaintyQuantification.generatePossibleStochasticParameters","page":"Library","title":"OpenGeoSysUncertaintyQuantification.generatePossibleStochasticParameters","text":"generatePossibleStochasticParameters(\n\tprojectfile::String,\n\tfile::String=\"./PossibleStochasticParameters.xml\",\n\tkeywords::Vector{String}=ogs_numeric_keyvals\n\t)\n\nHelper function for initial setup of a stochastic project. Scans an existing OGS6 projectfile (.prj) for keywords indicating numeric values usable as stochastic input parameter. Returns a Vector of Strings with OGS6 pathes. Uses OGS6 Pathes for parameter indentification (see Ogs6InputfileHandler.getAllPathesbyTag and Ogs6InputFileHandler.getElementbyPath). Outputs an XML file written by XMLParser.Julia2XML.\t\n\nArguments\n\nprojectfile::String: Path to OGS6 project file (e.g. ./path/to/project.prj)\nfile::String: Path of output XML file which gets written by this function.\nkeywords::Vector{String}: List of keywords.\n\n\n\n\n\n","category":"function"},{"location":"lib/lib/#OpenGeoSysUncertaintyQuantification.loadStochasticParameters","page":"Library","title":"OpenGeoSysUncertaintyQuantification.loadStochasticParameters","text":"loadStochasticParameters(file::String=\"./PossibleStochasticParameters.xml\")\n\nLoads an XML file containing OGS6 pathes by XMLParser.XML2Julia.\n\nArguments\n\nfile::String: Path of XML file with OGS6 pathes.\n\n\n\n\n\n","category":"function"},{"location":"lib/lib/#OpenGeoSysUncertaintyQuantification.generateStochasticOGSModell","page":"Library","title":"OpenGeoSysUncertaintyQuantification.generateStochasticOGSModell","text":"generateStochasticOGSModell(\nprojectfile::
[…]
