### 2024-05-30 c104866
update paper and docs
