+++ paper.md
@@ -28,9 +28,9 @@ In the simulation-aided design of safety-related projects, the effects of uncert
 
 # Statement of need
 
-Stability verifications for large structures can often only be carried out with the help of numerical simulations. A particularly difficult example is the safe storage of highly radioactive waste in subsurface repositories. These are usually planned at depths of several hundred meters. The heat radiated by the fuel rods can influence thermal-hydraulic-mechanical processes down to depths of several kilometers. Numerical simulations are needed to ensure site safety. Because of the outstanding interest in the safety of these repositories, the uncertainties are also given special attention. OpenGeoSys is a frequently used simulation tool in the German research community dedicated to the containment-safe storage of radioactive waste. Due to legal requirements in the repository safety ordinance, it is necessary to quantify input uncertainties of a deterministic OGS6 model.
+Stability verifications for large structures can often only be carried out with the help of numerical simulations. A particularly difficult example is the safe storage of highly radioactive waste in subsurface repositories. These are usually planned at depths of several hundred meters. The heat radiated by the fuel rods can influence thermal-hydraulic-mechanical processes down to depths of several kilometers. Numerical simulations are needed to ensure site safety [@kohl2023geolab]. Because of the outstanding interest in the safety of these repositories, the uncertainties are also given special attention [@kurgyis2024uncertainties]. OpenGeoSys is a frequently used simulation tool in the German research community dedicated to the containment-safe storage of radioactive waste. Due to legal requirements in the repository safety ordinance, it is necessary to quantify input uncertainties of a deterministic OGS6 model.
 
-Parameterizing a THM model in OpenGeoSys [@kolditz2012opengeosys] is complex and time-consuming due to the large number of possible input parameters. For easier applicability, this package was not designed as a generic uncertainty quantification framework but to relate it explicitly to OGS6. Nevertheless, care was taken to ensure that the underlying packages were as generic as possible and could be used in other projects.
+Parameterizing a THM model in OpenGeoSys [@kolditz2012opengeosys] is complex and time-consuming due to the large number of possible input parameters. This package was designed with the aim to be explicitly related to the OGS6 project. However, care was taken to ensure that the underlying packages could potentially be still applicable to other projects. 
 
 For this toolbox, the following stochastic modeling strategy was chosen: the less known the effect of the input uncertainty on the output, the more general the quantity of interest should be selected. As a first step, selecting a complete OGS6 postprocessing result is a viable option. This approach provides methods to `explore` uncertainties in the OGS6 simulation output.
 
