+++ paper.bib
@@ -119,3 +119,24 @@
   doi          = {10.5281/zenodo.2647458},
   url          = {https://doi.org/10.5281/zenodo.2647458}
 }
+
+@article{kohl2023geolab,
+  title={GeoLaB--Geothermal Laboratory in the crystalline Basement: synergies with research for a nuclear waste repository},
+  author={Kohl, Thomas and Sass, Ingo and Kolditz, Olaf and Sch{\"u}th, Christoph and R{\"u}haak, Wolfram and Schamp, J{\"u}rgen and Bremer, Judith and Rudolph, Bastian and Sch{\"a}tzler, Katharina and Schill, Eva},
+  journal={Safety of Nuclear Waste Disposal},
+  volume={2},
+  pages={135--136},
+  year={2023},
+  publisher={Copernicus Publications G{\"o}ttingen, Germany}
+}
+
+@article{kurgyis2024uncertainties,
+  title={Uncertainties and robustness with regard to the safety of a repository for high-level radioactive waste: introduction of a research initiative},
+  author={Kurgyis, Kata and Achtziger-Zupan{\v{c}}i{\v{c}}, Peter and Bjorge, Merle and Boxberg, Marc S and Broggi, Matteo and Buchwald, J{\"o}rg and Ernst, Oliver G and Fl{\"u}gge, Judith and Ganopolski, Andrey and Graf, Thomas and others},
+  journal={Environmental Earth Sciences},
+  volume={83},
+  number={2},
+  pages={82},
+  year={2024},
+  publisher={Springer}
+}
