### 2024-05-30 baxmittens/OpenGeoSysUncertaintyQuantification.jl#3 @baxmittens
Hello @ziyiyin97

thank you for your comments.
I gladly adopted your improvement suggestion from your first two points.

Regarding 3: I think since all of them are repositories which I own, it is comprehensible that I am the main or even only contributor to all this projects. For 2 of them there is a reference for a JOSS paper given with my name to it. Should be ok, I think or am I wrong?
