+++ test/runtests.jl
@@ -4,7 +4,11 @@ using Test
 @info "Running tests"
 
 @testset "Project setup" begin
-    include("test_project_setup.jl")
+    include("ex1/generate_stoch_params_file.jl")
+end
+
+@testset "Stochastic model setup" begin
+    include("ex1/generate_stoch_params_file.jl")
 end
 
 @testset "Utils" begin
@@ -12,8 +16,6 @@ end
 end
 
 @testset "OGS run" begin
-    include("ex1/generate_stoch_params_file.jl")
-    include("ex1/generate_stoch_model_lowres.jl")
     include("ex1/start.jl")
     @test maximum(varval["temperature_interpolated"][:,end]) > 1e7 && maximum(varval["temperature_interpolated"][:,end]) < 1e8
 end
