+++ test/test_project_setup.jl
@@ -1,5 +1,5 @@
-file_stoch_model = "./ex1/StochasticOGSModelParams.xml"
-file_sample_params = "./ex1/SampleMethodParams.xml"
+file_stoch_model = joinpath(TESTDIR, "StochasticOGSModelParams.xml")
+file_sample_params = joinpath(TESTDIR, "SampleMethodParams.xml")
 
 ogsuqparams = OGSUQParams(file_stoch_model, file_sample_params)
 
