+++ test/Examples/ASG_Point_Heat_Source/generate_stoch_model_lowres.jl
@@ -10,8 +10,8 @@ additionalprojecfilespath=joinpath(PATH,"mesh")
 outputpath=joinpath(PATH,"Res")
 
 # hierarchical sparse grid level
-INIT_LVL = 3
-MAX_LVL = 4
+INIT_LVL = 2
+MAX_LVL = 2
 
 #simcall = "ogs" # ogs binary is in path, otherwise put your path/to/bin/ogs here
 simcall = OpenGeoSysUncertaintyQuantification.install_ogs()
