+++ test/Sensitivity/Morris/generate_stoch_params_file.jl
@@ -1,4 +1,0 @@
-using OpenGeoSysUncertaintyQuantification
-
-projectfile="./project/point_heat_source_2D.prj"
-pathes = generatePossibleStochasticParameters(projectfile)
