+++ test/Examples/ASG_Point_Heat_Source/generate_stoch_params_file.jl
@@ -4,7 +4,7 @@ PATH = joinpath(splitpath(@__FILE__)[1:end-1]...)
 projectfile = joinpath(PATH, "project", "point_heat_source_2D.prj")
 output_xml = joinpath(PATH, "StochasticParameters.xml")
 
-pathes = generatePossibleStochasticParameters(projectfile)
+pathes = generatePossibleStochasticParameters(projectfile, output_xml)
 
 porosity_ind = findfirst(x->contains(x,"@id/0") && contains(x,"porosity"), pathes)
 liquid_th_cond_ind = findfirst(x->contains(x,"@id/0") && contains(x,"AqueousLiquid") && contains(x,"thermal_conductivity"), pathes)
