### 2024-05-30 #6725 @baxmittens
> I don't really see a good solution here; I don't want to get too involved in the nitty-gritty software design, but is there any way you can extract the binary interface such that you can test that it has the correct behavior with output from a dummy OGS process (i.e., record OGS output from a "reference" session that you know works and somehow use that to test the behavior of how OGSUQ.jl responds)? I understand if this ends up not being viable, it's just the main point of contention when finishing the checklist.

Hi @dannys4

I think, I finally fixed this issue.

You can now install the newest version of OpenGeoSysUncertaintyQuantification

```julia
]add OpenGeoSysUncertaintyQuantification@0.1.4
```

(don't forget to free other packages like the DistributedSparseGrid, like I did...)

I added a `OpenGeoSysUncertaintyQuantification.install_ogs()` function, which runs a shell script which installs OGS via pip3 ([see OGS docs](https://www.opengeosys.org/docs/userguide/basics/introduction/#install-via-pip)) in your `test` folder. However, you have to have a python version < 3.12 installed (i.e. linked with `python3` and `pip3` which are called by the script).
If you do
```julia
]test OpenGeoSysUncertaintyQuantification
```

the example, now located in `test/Examples/ASG_Point_Heat_Source` is copied to a temp folder in `test` and a reduced sampling with less collocation points is performed. I included a Github Action which performs the `runtests.jl` on each push and pull request. This works just fine.

You can also navigate to `test/Examples/ASG_Point_Heat_Source' and run

```julia
include("run_ex1.jl")
```

to run the full example. If you want to generate the plots, you have to comment in the last 10 lines or so in [`start.jl`](https://github.com/baxmittens/OpenGeoSysUncertaintyQuantification.jl/blob/main/test/Examples/ASG_Point_Heat_Source/start.jl) and you have to have `PlotlyJS` and `DistributedSparseGridsPlotting` installed.

I altered the documentation, accordingly. I tested this on 2 different machines. For me it worked both times.
