+++ paper.bib
@@ -120,13 +120,13 @@
   url          = {https://doi.org/10.5281/zenodo.2647458}
 }
 
-@article{kohl2023geolab,
-  title={GeoLaB--Geothermal Laboratory in the crystalline Basement: synergies with research for a nuclear waste repository},
-  author={Kohl, Thomas and Sass, Ingo and Kolditz, Olaf and Sch{\"u}th, Christoph and R{\"u}haak, Wolfram and Schamp, J{\"u}rgen and Bremer, Judith and Rudolph, Bastian and Sch{\"a}tzler, Katharina and Schill, Eva},
-  journal={Safety of Nuclear Waste Disposal},
-  volume={2},
-  pages={135--136},
-  year={2023},
+@article{hoyer2021preliminary,
+  title={Preliminary safety analyses in the high-level radioactive waste site selection procedure in Germany},
+  author={Hoyer, Eva-Maria and Luijendijk, Elco and M{\"u}ller, Paulina and Kreye, Phillip and Panitz, Florian and Gawletta, Dennis and R{\"u}haak, Wolfram},
+  journal={Advances in Geosciences},
+  volume={56},
+  pages={67--75},
+  year={2021},
   publisher={Copernicus Publications G{\"o}ttingen, Germany}
 }
 
