+++ paper.md
@@ -28,7 +28,7 @@ In the simulation-aided design of safety-related projects, the effects of uncert
 
 # Statement of need
 
-Stability verifications for large structures can often only be carried out with the help of numerical simulations. A particularly difficult example is the safe storage of highly radioactive waste in subsurface repositories. These are usually planned at depths of several hundred meters. The heat radiated by the fuel rods can influence thermal-hydraulic-mechanical processes down to depths of several kilometers. Numerical simulations are needed to ensure site safety [@kohl2023geolab]. Because of the outstanding interest in the safety of these repositories, the uncertainties are also given special attention [@kurgyis2024uncertainties]. OpenGeoSys is a frequently used simulation tool in the German research community dedicated to the containment-safe storage of radioactive waste. Due to legal requirements in the repository safety ordinance, it is necessary to quantify input uncertainties of a deterministic OGS6 model.
+Stability verifications for large structures can often only be carried out with the help of numerical simulations. A particularly difficult example is the safe storage of highly radioactive waste in subsurface repositories. These are usually planned at depths of several hundred meters. The heat radiated by the fuel rods can influence thermal-hydraulic-mechanical processes down to depths of several kilometers. Numerical simulations are needed to ensure site safety [@hoyer2021preliminary]. Because of the outstanding interest in the safety of these repositories, the uncertainties are also given special attention [@kurgyis2024uncertainties]. OpenGeoSys is a frequently used simulation tool in the German research community dedicated to the containment-safe storage of radioactive waste. Due to legal requirements in the repository safety ordinance, it is necessary to quantify input uncertainties of a deterministic OGS6 model.
 
 Parameterizing a THM model in OpenGeoSys [@kolditz2012opengeosys] is complex and time-consuming due to the large number of possible input parameters. This package was designed with the aim to be explicitly related to the OGS6 project. However, care was taken to ensure that the underlying packages could potentially be still applicable to other projects. 
 
