### 2024-05-30 #6725 @baxmittens
> to run the full example. If you want to generate the plots, you have to comment in the last 10 lines or so in [`start.jl`](https://github.com/baxmittens/OpenGeoSysUncertaintyQuantification.jl/blob/main/test/Examples/ASG_Point_Heat_Source/start.jl) and you have to have `PlotlyJS` and `DistributedSparseGridsPlotting` installed.

Before running the full example it is beneficial to increase the number of workers `n_workers` in the [`generate_stoch_model.jl .jl`](https://github.com/baxmittens/OpenGeoSysUncertaintyQuantification.jl/blob/554ac05ee44620b3e456647a4fdaed41d228fe78/test/Examples/ASG_Point_Heat_Source/generate_stoch_model.jl#L22) to the numbers of cores of your machine.
