+++ test/runtests.jl
@@ -4,14 +4,16 @@ using Test
 @info "Running tests"
 
 TESTDIR = "./tmp_test_dir"
-EXAMPLEDIR = "./Examples/ASG_Point_Heat_Source/"
+EXAMPLEDIR = "./Examples/OGSUQ/ASG_Consolidation_PHS/"
 
 if isdir(TESTDIR)
     @info "deleting $TESTDIR"
     run(`rm -rf $TESTDIR`)
 end
-
-cp(EXAMPLEDIR, TESTDIR)
+mkdir(TESTDIR)
+cp(joinpath(EXAMPLEDIR, "generate_stoch_params_file.jl"), joinpath(TESTDIR, "generate_stoch_params_file.jl"))
+cp(joinpath(EXAMPLEDIR, "generate_stoch_model_lowres.jl"), joinpath(TESTDIR,"generate_stoch_model_lowres.jl"))
+cp(joinpath(EXAMPLEDIR, "start.jl"), joinpath(TESTDIR,"start.jl"))
 
 @testset "Stochastic parameters" begin
     include(joinpath(TESTDIR, "generate_stoch_params_file.jl"))
