+++ test/Examples/OGSUQ/ASG_Sat_Mass_Transport/run_all.jl
@@ -1,0 +1,3 @@
+include("generate_stoch_params_file.jl")
+include("generate_stoch_model.jl")
+include("start.jl")

[…]
