+++ test/Examples/OGSUQ/MC_Consolidation_PHS/generate_stoch_model.jl
@@ -10,9 +10,9 @@ stoch_params_xml = joinpath(PATH, "StochasticParameters.xml")
 samplemethod_output_xml = joinpath(PATH, "SampleMethodParams.xml")
 outputpath=joinpath(PATH,"Res")
 
-simcall = "/Users/maximilianbittens/Documents/GitHub/OpenGeoSys/build/release/bin/ogs"
+#simcall = "/Users/maximilianbittens/Documents/GitHub/OpenGeoSys/build/release/bin/ogs"
 #simcall = "ogs" # ogs binary is in path, otherwise put your path/to/bin/ogs here
-#simcall = OpenGeoSysUncertaintyQuantification.install_ogs()
+simcall = OpenGeoSysUncertaintyQuantification.install_ogs()
 @info "simcall: $simcall"
 
 postprocfiles=["PointHeatSource_quarter_002_2nd.xdmf"]
