+++ test/Examples/OGSUQ/ASG_Sat_Mass_Transport/generate_stoch_model.jl
@@ -11,17 +11,17 @@ samplemethod_output_xml = joinpath(PATH, "SampleMethodParams.xml")
 outputpath=joinpath(PATH,"Res")
 
 # hierarchical sparse grid level
-INIT_LVL = 4
-MAX_LVL = 12
+INIT_LVL = 6
+MAX_LVL = 14
 
-simcall = "/Users/maximilianbittens/Documents/GitHub/OpenGeoSys/build/release/bin/ogs"
+#simcall = "/Users/maximilianbittens/Documents/GitHub/OpenGeoSys/build/release/bin/ogs"
 #simcall = "ogs" # ogs binary is in path, otherwise put your path/to/bin/ogs here
-#simcall = OpenGeoSysUncertaintyQuantification.install_ogs()
+simcall = OpenGeoSysUncertaintyQuantification.install_ogs()
 @info "simcall: $simcall"
 
 postprocfiles=["DiffusionAndStorageAndAdvectionAndDispersionHalf_square_1x1_quad_1e3.xdmf"]
 stochmethod=AdaptiveHierarchicalSparseGrid
-n_workers = 21
+n_workers = 25
 
 stochparampathes = loadStochasticParameters(stoch_params_xml)
 
@@ -53,5 +53,5 @@ samplemethodparams = generateSampleMethodModel(stochasticmodelparams, samplemeth
 # alter sample method params
 samplemethodparams.init_lvl = INIT_LVL
 samplemethodparams.maxlvl = MAX_LVL
-samplemethodparams.tol = 0.005
+samplemethodparams.tol = 0.001
 write(samplemethodparams)
