### 2024-06-04 baxmittens/OpenGeoSysUncertaintyQuantification.jl#3 @ziyiyin97
In terms of the software and documentation, I am not sure how to run the examples in this section https://baxmittens.github.io/OpenGeoSysUncertaintyQuantification.jl/dev/#Usage. The path does not exist. It seems that this example is outdated (probably due to the changes requested by the other reviewer). Please fix. Thanks!
