+++ test/Examples/OGSUQ/Integration_2D/generate_stoch_model.jl
@@ -1,9 +1,9 @@
 using OpenGeoSysUncertaintyQuantification
 
 PATH = joinpath(splitpath(@__FILE__)[1:end-1]...)
-OGS_PRJ_PATH = joinpath(OpenGeoSysUncertaintyQuantification.ogs_prj_folder(), "Consolidation_PHS")
-projectfile= joinpath(OGS_PRJ_PATH,"prj","point_heat_source_2D.prj")
-additionalprojecfilespath=joinpath(OGS_PRJ_PATH,"mesh")
+OGS_PRJ_PATH = joinpath(OpenGeoSysUncertaintyQuantification.ogs_prj_folder(), "Saturated_Mass_Transport")
+projectfile= joinpath(OGS_PRJ_PATH,"prj","DiffusionAndStorageAndAdvectionAndDispersionHalf.prj")
+additionalprojecfilespath=joinpath(OGS_PRJ_PATH,"misc")
 user_functions_file = joinpath(PATH, "user_functions.jl")
 output_xml = joinpath(PATH, "StochasticOGSModelParams.xml")
 stoch_params_xml = joinpath(PATH, "StochasticParameters.xml")
@@ -19,7 +19,7 @@ simcall = "/Users/maximilianbittens/Documents/GitHub/OpenGeoSys/build/release/bi
 #simcall = OpenGeoSysUncertaintyQuantification.install_ogs()
 @info "simcall: $simcall"
 
-postprocfiles=["PointHeatSource_quarter_002_2nd.xdmf"]
+postprocfiles=["DiffusionAndStorageAndAdvectionAndDispersionHalf_square_1x1_quad_1e3.xdmf"]
 stochmethod=AdaptiveHierarchicalSparseGrid
 n_workers = 25
 
@@ -40,13 +40,12 @@ stochasticmodelparams = generateStochasticOGSModell(
 
 # alter the stochastic parameters
 stoch_params = stoch_parameters(stochasticmodelparams)
-@assert contains(stoch_params[1].path, "AqueousLiquid")
-stoch_params[1].dist = Normal(0.6,0.175)
-stoch_params[1].lower_bound = 0.3
-stoch_params[1].upper_bound = 0.9
-stoch_params[2].dist = Normal(0.45,0.15)
-stoch_params[2].lower_bound = 0.1
-stoch_params[2].upper_bound = 0.8
+stoch_params[1].dist = Normal(0.55,0.25)
+stoch_params[1].lower_bound = 0.1
+stoch_params[1].upper_bound = 1.0
+stoch_params[2].dist = Normal(0.65,0.35)
+stoch_params[2].lower_bound = 0.2
+stoch_params[2].upper_bound = 1.4
 write(stochasticmodelparams)
 
 #generate sample method model
