+++ test/Examples/OGSUQ/ASG_Consolidation_PHS/generate_stoch_model.jl
@@ -14,9 +14,9 @@ outputpath=joinpath(PATH,"Res")
 INIT_LVL = 6
 MAX_LVL = 14
 
-simcall = "/Users/maximilianbittens/Documents/GitHub/OpenGeoSys/build/release/bin/ogs"
+#simcall = "/Users/maximilianbittens/Documents/GitHub/OpenGeoSys/build/release/bin/ogs"
 #simcall = "ogs" # ogs binary is in path, otherwise put your path/to/bin/ogs here
-#simcall = OpenGeoSysUncertaintyQuantification.install_ogs()
+simcall = OpenGeoSysUncertaintyQuantification.install_ogs()
 @info "simcall: $simcall"
 
 postprocfiles=["PointHeatSource_quarter_002_2nd.xdmf"]
