### 2024-06-05 120ddc6
Update generate_stoch_model_lowres.jl
