+++ test/Examples/OGSUQ/ASG_Consolidation_PHS/generate_stoch_model_lowres.jl
@@ -1,12 +1,13 @@
 using OpenGeoSysUncertaintyQuantification
 
 PATH = joinpath(splitpath(@__FILE__)[1:end-1]...)
-projectfile= joinpath(PATH,"project","point_heat_source_2D.prj")
+OGS_PRJ_PATH = joinpath(OpenGeoSysUncertaintyQuantification.ogs_prj_folder(), "Consolidation_PHS")
+projectfile= joinpath(OGS_PRJ_PATH,"prj","point_heat_source_2D.prj")
+additionalprojecfilespath=joinpath(OGS_PRJ_PATH,"mesh")
 user_functions_file = joinpath(PATH, "user_functions.jl")
 output_xml = joinpath(PATH, "StochasticOGSModelParams.xml")
 stoch_params_xml = joinpath(PATH, "StochasticParameters.xml")
 samplemethod_output_xml = joinpath(PATH, "SampleMethodParams.xml")
-additionalprojecfilespath=joinpath(PATH,"mesh")
 outputpath=joinpath(PATH,"Res")
 
 # hierarchical sparse grid level
