### 2024-06-05 baxmittens/OpenGeoSysUncertaintyQuantification.jl#3 @baxmittens
@ziyiyin97
You're right. I altered the documentation.
You can update to OpenGeoSysUncertaintyQuantification@0.1.5 and the example and the test should work. Hope this fixes your issues. Otherwise, please write again here.
Thank you.
