### 2024-06-09 baxmittens/OpenGeoSysUncertaintyQuantification.jl#3 @ziyiyin97
Thanks! Things run fine now. I think this change is necessary https://github.com/baxmittens/OpenGeoSysUncertaintyQuantification.jl/pull/4

Also it would be good to mention to the users that running this complete example might take a long while (i.e. it is not a "minimal" example that one expects to finish running in a few seconds). Also would be good to mention if the users need a reasonable machine to run this example (i.e. can users run this example on a small laptop given memory or other constraints?)
