### 2024-06-10 baxmittens/OpenGeoSysUncertaintyQuantification.jl#3 @baxmittens
Hello @ziyiyin97
I merged your pull request. Thank you for that catch.
In the next release, the test will only generate the bare minimum of sample points.
![image](https://github.com/baxmittens/OpenGeoSysUncertaintyQuantification.jl/assets/100423479/87506e5e-a479-4ca2-8e44-37e5de404e23)
For me the tests runs in under one minute which is fine, is guess. Right?

Greetz max
