### 2024-06-14 #6725 @dannys4
Just finished my checklist as well. The tests work when I download it from the general registry and the examples also seem to work (I had to use the low-res version; thanks for including that!). The new examples do a great job of laying out exactly what needs to happen and (I think) will greatly benefit users. @kanishkan91 Let me know if anything else needs to get addressed on my end; otherwise, it looks great and I look forward to the paper.
