### 2024-06-14 #6725 @kanishkan91
👋 @baxmittens  - I went through the paper one more time and we are almost there!  A few clean ups may be necessary. It seems no DOIs could be found for 4 references (See above).

After that is just setting up the archive for your new release.

So here is what we have left to do:

- [ ] Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list matches the people and order in your paper). You may also add the authors' ORCID.

- [ ] Please respond with the 1) version and 2) DOI of the archived version here
