+++ paper.bib
@@ -95,7 +95,8 @@
   volume={67},
   pages={589--599},
   year={2012},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s12665-012-1546-x}
 }
 
 @misc{Distributions.jl-2019,
@@ -127,7 +128,8 @@
   volume={56},
   pages={67--75},
   year={2021},
-  publisher={Copernicus Publications G{\"o}ttingen, Germany}
+  publisher={Copernicus Publications G{\"o}ttingen, Germany},
+  doi={10.5194/adgeo-56-67-2021}
 }
 
 @article{kurgyis2024uncertainties,
@@ -138,5 +140,6 @@
   number={2},
   pages={82},
   year={2024},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s12665-023-11346-8}
 }
