+++ paper.bib
@@ -1,51 +1,3 @@
-@inproceedings{smolyak1963quadrature,
-  title={Quadrature and interpolation formulas for tensor products of certain classes of functions},
-  author={Smolyak, Sergei Abramovich},
-  booktitle={Doklady Akademii Nauk},
-  volume={148},
-  number={5},
-  pages={1042--1045},
-  year={1963},
-  organization={Russian Academy of Sciences}
-}
-
-@article{babuvska2007stochastic,
-  title={A stochastic collocation method for elliptic partial differential equations with random input data},
-  author={Babu{\v{s}}ka, Ivo and Nobile, Fabio and Tempone, Ra{\'u}l},
-  journal={SIAM Journal on Numerical Analysis},
-  volume={45},
-  number={3},
-  pages={1005--1034},
-  year={2007},
-  publisher={SIAM},
-  doi = {10.1137/100786356}
-}
-
-
-@article{nobile2008sparse,
-  title={A sparse grid stochastic collocation method for partial differential equations with random input data},
-  author={Nobile, Fabio and Tempone, Ra{\'u}l and Webster, Clayton G},
-  journal={SIAM Journal on Numerical Analysis},
-  volume={46},
-  number={5},
-  pages={2309--2345},
-  year={2008},
-  publisher={SIAM},
-  doi = {10.1137/060663660}
-}
-
-@article{ma2009adaptive,
-  title={An adaptive hierarchical sparse grid collocation algorithm for the solution of stochastic differential equations},
-  author={Ma, Xiang and Zabaras, Nicholas},
-  journal={Journal of Computational Physics},
-  volume={228},
-  number={8},
-  pages={3084--3113},
-  year={2009},
-  publisher={Elsevier},
-  doi = {10.1016/j.jcp.2009.01.006}
-}
-
 @article{gates2015multilevel,
   title={A multilevel adaptive sparse grid stochastic collocation approach to the non-smooth forward propagation of uncertainty in discretized problems},
   author={Gates, Robert L and Bittens, Maximilian R},
@@ -63,18 +15,6 @@
   pages={4300},
   year={2022},
   doi = {10.21105/joss.04300}
-}
-
-@article{schroeder2000visualizing,
-  title={Visualizing with VTK: a tutorial},
-  author={Schroeder, William J and Avila, Lisa Sobierajski and Hoffman, William},
-  journal={IEEE Computer graphics and applications},
-  volume={20},
-  number={5},
-  pages={20--27},
-  year={2000},
-  publisher={IEEE},
-  doi={10.1109/38.865875}
 }
 
 @article{bittens2023distributedsparsegrids,
