### 2024-06-17 #6725 @baxmittens
Hi there,

I've added the DOIs of the citations, released a new version and made a Zenodo archive from that.

https://github.com/baxmittens/OpenGeoSysUncertaintyQuantification.jl/releases/tag/v0.1.6
https://zenodo.org/records/11923517

That should be it, right?

Greetz Max
