### 2024-06-17 #6725 @kanishkan91
@baxmittens I have recommended this for acceptance now. The AEiC in this submission track will review shortly and if all goes well this will go live soon! Big thank you to @dannys4  and @ziyiyin97 for reviewing! JOSS is volunteer run and relies heavily on researchers such as yourself.
