### 2024-06-17 #6725 @kthyng
@baxmittens In your paper please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, "julia" and "germany" are not capitalized; please check for others.
