+++ paper.bib
@@ -7,7 +7,7 @@
 }
 
 @article{bittens2022vtufilehandler,
-  title={VTUFileHandler: A VTU library in the {J}ulia language that implements an algebra for basic mathematical operations on VTU data},
+  title={VTUFileHandler: A {VTU} library in the {J}ulia language that implements an algebra for basic mathematical operations on {VTU} data},
   author={Bittens, Maximilian},
   journal={Journal of Open Source Software},
   volume={7},
@@ -18,7 +18,7 @@
 }
 
 @article{bittens2023distributedsparsegrids,
-  title={DistributedSparseGrids. jl: A Julia library implementing an Adaptive Sparse Grid collocation method},
+  title={DistributedSparseGrids. jl: A {J}ulia library implementing an {A}daptive {S}parse {G}rid collocation method},
   author={Bittens, Maximilian and Gates, Robert L},
   journal={Journal of Open Source Software},
   volume={8},
@@ -29,7 +29,7 @@
 }
 
 @article{kolditz2012opengeosys,
-  title={OpenGeoSys: an open-source initiative for numerical simulation of thermo-hydro-mechanical/chemical (THM/C) processes in porous media},
+  title={Open{G}eo{S}ys: an open-source initiative for numerical simulation of thermo-hydro-mechanical/chemical ({THM/C}) processes in porous media},
   author={Kolditz, Olaf and Bauer, Sebastian and Bilke, Lars and B{\"o}ttcher, Norbert and Delfs, Jens-Olaf and Fischer, Thomas and G{\"o}rke, Uwe J and Kalbacher, Thomas and Kosakowski, Georg and McDermott, CI and others},
   journal={Environmental Earth Sciences},
   volume={67},
@@ -62,7 +62,7 @@
 }
 
 @article{hoyer2021preliminary,
-  title={Preliminary safety analyses in the high-level radioactive waste site selection procedure in Germany},
+  title={Preliminary safety analyses in the high-level radioactive waste site selection procedure in {G}ermany},
   author={Hoyer, Eva-Maria and Luijendijk, Elco and M{\"u}ller, Paulina and Kreye, Phillip and Panitz, Florian and Gawletta, Dennis and R{\"u}haak, Wolfram},
   journal={Advances in Geosciences},
   volume={56},
