### 2024-06-20 #6725 @kthyng
Congratulations on your new publication @baxmittens! Many thanks to @kanishkan91 and to reviewers @ziyiyin97 and @dannys4 for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
