### 2024-06-20 #6725 @kthyng
@baxmittens If you're interested in joining JOSS as a reviewer, please register at https://reviewers.joss.theoj.org/.
