### 2024-06-20 #6725 @baxmittens
Once again, thank you all for reviewing my software project.

@kthyng I recently did a review in JOSS, and I am willing to continue doing reviews. I completed my profile. JOSS is an excellent journal!
